backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-telegram-bot[webhooks]>=21.0
apscheduler
openpyxl
python-multipart
anthropic
xlrd
